pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gitguard-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "gitpython>=3.1.45",
    "google-genai>=0.3.0",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "typer>=0.21.0",
    "pydantic>=2.0.0",
]

[project.scripts]
gitguard = "gitguard.main:app"
